requirements.txt
numpy
scipy
sounddevice
PyYAML
PyQt6
